compilation warninf

+++ 2-selection_sort.c
@@ -6,7 +6,7 @@
  */
 void selection_sort(int *array, size_t size)
 {
-    int i, j, min_idx, tmp;
+    size_t i, j, min_idx, tmp;
   
 
     for (i = 0; i < size - 1; i++)
